+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/airflow.rb
@@ -128,6 +128,10 @@ class Airflow
     return (1.0 / 150.0).round(6) # Table 4.2.2(1) - Crawlspaces
   end
 
+  def self.get_default_unvented_space_ach()
+    return 0.1 # Assumption
+  end
+
   def self.get_default_mech_vent_fan_power(vent_fan)
     # 301-2019: Table 4.2.2(1b)
     # Returns fan power in W/cfm
@@ -1116,7 +1120,7 @@ class Airflow
 
     space = @spaces[HPXML::LocationBasementUnconditioned]
     volume = UnitConversions.convert(space.volume, 'm^3', 'ft^3')
-    ach = 0.1 # Assumption
+    ach = get_default_unvented_space_ach()
     cfm = ach / UnitConversions.convert(1.0, 'hr', 'min') * volume
     apply_infiltration_to_unconditioned_space(model, space, ach, nil, nil, nil)
   end
@@ -1126,8 +1130,9 @@ class Airflow
 
     space = @spaces[HPXML::LocationCrawlspaceVented]
     volume = UnitConversions.convert(space.volume, 'm^3', 'ft^3')
+    height = Geometry.get_height_of_spaces([space])
     sla = vented_crawl.vented_crawlspace_sla
-    ach = get_infiltration_ACH_from_SLA(sla, 8.202, weather)
+    ach = get_infiltration_ACH_from_SLA(sla, height, weather)
     cfm = ach / UnitConversions.convert(1.0, 'hr', 'min') * volume
     apply_infiltration_to_unconditioned_space(model, space, ach, nil, nil, nil)
   end
@@ -1137,8 +1142,7 @@ class Airflow
 
     space = @spaces[HPXML::LocationCrawlspaceUnvented]
     volume = UnitConversions.convert(space.volume, 'm^3', 'ft^3')
-    sla = 0 # Assumption
-    ach = get_infiltration_ACH_from_SLA(sla, 8.202, weather)
+    ach = get_default_unvented_space_ach()
     cfm = ach / UnitConversions.convert(1.0, 'hr', 'min') * volume
     apply_infiltration_to_unconditioned_space(model, space, ach, nil, nil, nil)
   end
@@ -1183,16 +1187,10 @@ class Airflow
     return if @spaces[HPXML::LocationAtticUnvented].nil?
 
     space = @spaces[HPXML::LocationAtticUnvented]
-    area = UnitConversions.convert(space.floorArea, 'm^2', 'ft^2')
     volume = UnitConversions.convert(space.volume, 'm^3', 'ft^3')
-    hor_lk_frac = 0.75
-    neutral_level = 0.5
-    sla = 0 # Assumption
-    ach = get_infiltration_ACH_from_SLA(sla, 8.202, weather)
-    ela = sla * area
+    ach = get_default_unvented_space_ach()
     cfm = ach / UnitConversions.convert(1.0, 'hr', 'min') * volume
-    c_w_SG, c_s_SG = calc_wind_stack_coeffs(site, hor_lk_frac, neutral_level, space)
-    apply_infiltration_to_unconditioned_space(model, space, nil, ela, c_w_SG, c_s_SG)
+    apply_infiltration_to_unconditioned_space(model, space, ach, nil, nil, nil)
   end
 
   def self.apply_local_ventilation(model, vent_object_array, obj_type_name)