+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/minitest_helper.rb
@@ -1,8 +1,22 @@
 # frozen_string_literal: true
 
-require 'minitest/autorun'
+$VERBOSE = nil # Prevents ruby warnings, see https://github.com/NREL/OpenStudio/issues/4301
 
+called_from_cli = true
+begin
+  OpenStudio.getOpenStudioCLI
+rescue
+  called_from_cli = false
+end
+
+if ENV['CI'] && !called_from_cli
+  require 'simplecov'
+  SimpleCov.coverage_dir(File.join(Dir.getwd, 'coverage'))
+  SimpleCov.formatter = SimpleCov::Formatter::HTMLFormatter
+  SimpleCov.start
+end
+
+require 'minitest/autorun'
 require 'minitest/reporters'
 require 'minitest/reporters/spec_reporter' # Needed when run via OS CLI
-
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new # spec-like progress